requirements.txt
docopt==0.6.2
pyserial==3.4
fastapi==0.54.1
uvicorn==0.7.0
